GH-180: [refactor] Change attendance to an enum
