+++ Microservices/event-service/src/main/java/app/sportahub/eventservice/model/event/participant/ParticipantAttendStatus.java
@@ -1,0 +1,7 @@
+package app.sportahub.eventservice.model.event.participant;
+
+public enum ParticipantAttendStatus {
+    JOINED,
+    CONFIRMED,
+    CANCELLED
+}
